Refactorizar código
