+++ views/board.pug
